refactor: 상품 DTO 범위 제한
